Restructure consumer and producer into own files
