+++ instrumentation/github.com/confluentinc/confluent-kafka-go/kafka/splunkkafka/consumer.go
@@ -15,7 +15,7 @@
 //go:build !windows
 // +build !windows
 
-package splunkkafka // import "github.com/signalfx/splunk-otel-go/instrumentation/github.com/confluentinc/confluent-kafka-go/kafka/splunkkafka"
+package splunkkafka
 
 import (
 	"context"
@@ -24,10 +24,18 @@ import (
 
 	"github.com/confluentinc/confluent-kafka-go/kafka"
 	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/codes"
 	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
 	"go.opentelemetry.io/otel/trace"
 )
+
+// Consumer wraps a kafka.Consumer and traces its operations.
+type Consumer struct {
+	*kafka.Consumer
+	cfg    config
+	group  string
+	events chan kafka.Event
+	prev   trace.Span
+}
 
 // NewConsumer calls kafka.NewConsumer and wraps the resulting Consumer with
 // tracing instrumentation.
@@ -49,25 +57,6 @@ func NewConsumer(conf *kafka.ConfigMap, opts ...Option) (*Consumer, error) {
 	}
 	wrapped.events = wrapped.traceEventsChannel(c.Events())
 	return wrapped, nil
-}
-
-// NewProducer calls kafka.NewProducer and wraps the resulting Producer with
-// tracing instrumentation.
-func NewProducer(conf *kafka.ConfigMap, opts ...Option) (*Producer, error) {
-	p, err := kafka.NewProducer(conf)
-	if err != nil {
-		return nil, err
-	}
-	return WrapProducer(p, opts...), nil
-}
-
-// Consumer wraps a kafka.Consumer and traces its operations.
-type Consumer struct {
-	*kafka.Consumer
-	cfg    config
-	group  string
-	events chan kafka.Event
-	prev   trace.Span
 }
 
 // WrapConsumer wraps a kafka.Consumer so that any consumed events are traced.
@@ -174,112 +163,4 @@ func (c *Consumer) Poll(timeoutMS int) (event kafka.Event) {
 		c.prev = c.startSpan(msg)
 	}
 	return evt
-}
-
-// A Producer wraps a kafka.Producer and traces its operations.
-type Producer struct {
-	*kafka.Producer
-	cfg            config
-	produceChannel chan *kafka.Message
-}
-
-// WrapProducer wraps a kafka.Producer so that any produced events are traced.
-func WrapProducer(p *kafka.Producer, opts ...Option) *Producer {
-	wrapped := &Producer{
-		Producer: p,
-		cfg:      newConfig(opts...),
-	}
-	wrapped.produceChannel = wrapped.traceProduceChannel(p.ProduceChannel())
-	return wrapped
-}
-
-func (p *Producer) traceProduceChannel(out chan *kafka.Message) chan *kafka.Message {
-	if out == nil {
-		return nil
-	}
-
-	in := make(chan *kafka.Message, cap(out))
-	go func() {
-		for msg := range in {
-			span := p.startSpan(msg)
-			out <- msg
-			span.End()
-		}
-	}()
-
-	return in
-}
-
-func (p *Producer) startSpan(msg *kafka.Message) trace.Span {
-	carrier := NewMessageCarrier(msg)
-	psc := p.cfg.Propagator.Extract(context.Background(), carrier)
-
-	attrs := []attribute.KeyValue{
-		semconv.MessagingSystemKey.String("kafka"),
-		semconv.MessagingDestinationKindTopic,
-		semconv.MessagingDestinationKey.String(*msg.TopicPartition.Topic),
-		semconv.MessagingMessageIDKey.String(strconv.FormatInt(int64(msg.TopicPartition.Offset), 10)), // nolint: gomnd
-		semconv.MessagingKafkaMessageKeyKey.String(string(msg.Key)),
-		semconv.MessagingKafkaPartitionKey.Int64(int64(msg.TopicPartition.Partition)),
-	}
-	opts := []trace.SpanStartOption{
-		trace.WithAttributes(attrs...),
-		trace.WithSpanKind(trace.SpanKindProducer),
-	}
-
-	name := fmt.Sprintf("%s send", *msg.TopicPartition.Topic)
-	ctx, span := p.cfg.Tracer.Start(psc, name, opts...)
-	// Inject the current span into the original message so it can be used to
-	// propagate the span.
-	p.cfg.Propagator.Inject(ctx, carrier)
-	return span
-}
-
-// Close calls the wrapped Producer.Close and closes the producer channel.
-func (p *Producer) Close() {
-	close(p.produceChannel)
-	p.Producer.Close()
-}
-
-// Produce calls the wrapped Producer.Produce and traces the request.
-func (p *Producer) Produce(msg *kafka.Message, deliveryChan chan kafka.Event) error {
-	span := p.startSpan(msg)
-
-	// if the user has selected a delivery channel, we will wrap it and
-	// wait for the delivery event to finish the span
-	if deliveryChan != nil {
-		oldDeliveryChan := deliveryChan
-		deliveryChan = make(chan kafka.Event)
-		go func() {
-			var err error
-			evt := <-deliveryChan
-			if respMsg, ok := evt.(*kafka.Message); ok {
-				// delivery errors are returned via TopicPartition.Error
-				err = respMsg.TopicPartition.Error
-			}
-			if err != nil {
-				span.RecordError(err)
-				span.SetStatus(codes.Error, err.Error())
-			}
-			span.End()
-			oldDeliveryChan <- evt
-		}()
-	}
-
-	err := p.Producer.Produce(msg, deliveryChan)
-	// with no delivery channel, finish immediately
-	if deliveryChan == nil {
-		if err != nil {
-			span.RecordError(err)
-			span.SetStatus(codes.Error, err.Error())
-		}
-		span.End()
-	}
-
-	return err
-}
-
-// ProduceChannel returns the traced producer channel.
-func (p *Producer) ProduceChannel() chan *kafka.Message {
-	return p.produceChannel
 }
